requirements.txt
amqp==2.5.1
aniso8601==7.0.0
astroid==2.3.2

billiard==3.6.1.0
boto3==1.13.0

braintree==4.2.0
celery==4.3.0
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==3.3.1
Django==2.0.5
django-braces==1.13.0
django-cors-headers==3.1.1
django-currentuser==0.5.0
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-model-utils==3.2.0
django-ratings==0.3.7
django-star-ratings==0.8.0
django-storages==1.9.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
docutils==0.15.2
future==0.16.0
graphene==2.1.8
graphene-django==2.6.0
graphene-file-upload==1.2.2
graphene-permissions==1.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.7
importlib-metadata==0.23
isort==4.3.21
jmespath==0.10.0
kombu==4.6.5
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
pathspec==0.5.9
paypalrestsdk==1.13.1
Pillow==8.0.1
promise==2.2.1
psycopg2==2.8.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.3
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-resize-image==1.1.19
pytz==2019.2
PyYAML==5.3.1
redis==3.3.8
requests==2.20.1
Rx==1.6.1
s3transfer==0.3.3
semantic-version==2.5.0
singledispatch==3.4.0.3
six==1.12
sqlparse==0.3.0
swapper==1.1.1
termcolor==1.1.0
typed-ast==1.4.0

vine==1.3.0
wcwidth==0.1.9
wrapt==1.11.2
zipp==0.6.0
